requirements.txt
Django>=2.1.3 ,<2.3.0
djangorestframework>=3.10.0,<3.11.0
psycopg2>=2.8.0,<2.9.0
flake8>=3.7.0,<3.8.0 
